refactor: spring logging aop에 맞는 디렉토리 적용
